Community ERP/requirements.txt
# Community ERP System - Backend Dependencies

# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# MongoDB Database
pymongo==4.6.1

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Environment Variables
python-dotenv==1.0.0

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Image Processing and AI Models
torch==2.1.0
torchvision==0.16.0
transformers==4.36.0
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78

# PaddleOCR for Text Extraction
paddlepaddle==2.6.0
paddleocr==2.7.3

# Optional: Async MongoDB Driver
motor==3.3.2
